Fix regex for instance type and cve detection (#28)
* Fix regex for instance type and cve detection

Previously only the first alternative in these regex is tested if it matches the beginning of the test string and the
last alternative is tested if it matches the end of the test string. What we want to test is if any of the alternatives
match the test string. Hence the added grouping around all alternatives.

* Fix plugins.yaml test: job-dsl 1.83 has no cve